Move event module from src/components to src/
* Rename EngineEvent to WindowEvent
