+++ examples/03_renderable/main.rs
@@ -7,7 +7,7 @@ use amethyst::config::Element;
 use amethyst::ecs::{World, Join};
 use amethyst::gfx_device::DisplayConfig;
 use amethyst::asset_manager::AssetManager;
-use amethyst::components::event::EngineEvent;
+use amethyst::event::WindowEvent;
 use amethyst::renderer::{VertexPosNormal, Pipeline};
 
 use self::genmesh::generators::{SphereUV};
@@ -116,8 +116,8 @@ impl State for Example {
         Trans::None
     }
 
-    fn handle_events(&mut self, events: &[EngineEvent], _: &mut World, _: &mut AssetManager, _: &mut Pipeline) -> Trans {
-        use amethyst::components::event::*;
+    fn handle_events(&mut self, events: &[WindowEvent], _: &mut World, _: &mut AssetManager, _: &mut Pipeline) -> Trans {
+        use amethyst::event::*;
         for event in events {
             match event.payload {
                 Event::KeyboardInput(_, _, Some(VirtualKeyCode::Escape)) => return Trans::Quit,
