+++ src/components/mod.rs
@@ -2,4 +2,3 @@
 
 pub mod rendering;
 pub mod transform;
-pub mod event;
